+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Release 4.0.0
+
+### New features
+- Supports FlowConfig feature to link data provider to put data to FTP server
+- Provide version of module via 'OnNewStatusModuleVersion'
+- Check if features of module can be used on device and provide this via 'OnNewStatusModuleIsActive' event / 'getStatusModuleActive' function
+- Function 'getParameters' to provide PersistentData parameters
+- It is now possible to register to events to put received data (JPG or DATA) to the FTP server (see 'addRegistration')
+- Function to 'resetModule' to default setup
+
+### Improvements
+- New UI design available (e.g. selectable via CSK_Module_PersistentData v4.1.0 or higher), see 'OnNewStatusCSKStyle'
+- 'loadParameters' returns its success
+- 'sendParameters' can control if sent data should be saved directly by CSK_Module_PersistentData
+- Changed log level of some messages from 'info' to 'fine'
+- Added UI icon and browser tab information
+
 ## Release 3.0.0
 
 ### Improvements